Remove redundant session manager and session handler
