+++ toolkit/committee-selection/pallet/src/mock.rs
@@ -1,4 +1,7 @@
-use crate::{Call, pallet, session_manager::ValidatorManagementSessionManager};
+use crate::{
+	Call, pallet, pallet_session_support::PalletSessionSupport,
+	session_manager::ValidatorManagementSessionManager,
+};
 use frame_support::{
 	dispatch::PostDispatchInfo,
 	pallet_prelude::*,
@@ -150,7 +153,7 @@ impl pallet_session::Config for Test {
 	type ValidatorIdOf = sp_runtime::traits::ConvertInto;
 	type ShouldEndSession = pallet_session::PeriodicSessions<Period, Offset>;
 	type NextSessionRotation = pallet_session::PeriodicSessions<Period, Offset>;
-	type SessionManager = ValidatorManagementSessionManager<Test>;
+	type SessionManager = PalletSessionSupport<Test>;
 	type SessionHandler = TestSessionHandler;
 	type Keys = SessionKeys;
 	type DisablingStrategy = ();
