+++ toolkit/committee-selection/pallet/src/pallet_session_support.rs
@@ -4,9 +4,12 @@
 use crate::CommitteeMember;
 use core::marker::PhantomData;
 use derive_new::new;
+use frame_support::traits::UnfilteredDispatchable;
+use frame_system::RawOrigin;
 use frame_system::pallet_prelude::BlockNumberFor;
-use log::{debug, warn};
+use log::{debug, info, warn};
 use pallet_partner_chains_session::SessionIndex;
+use sp_std::collections::btree_set::BTreeSet;
 use sp_std::vec::Vec;
 
 /// Implements [pallet_session::SessionManager] and [pallet_session::ShouldEndSession] integrated with [crate::Pallet].
@@ -20,6 +23,8 @@ pub struct PalletSessionSupport<T> {
 
 impl<T: crate::Config + pallet_session::Config> pallet_session::SessionManager<T::AccountId>
 	for PalletSessionSupport<T>
+where
+	<T as pallet_session::Config>::Keys: From<T::AuthorityKeys>,
 {
 	/// Sets the first validator-set by mapping the current committee from [crate::Pallet]
 	fn new_session_genesis(_new_index: SessionIndex) -> Option<Vec<T::AccountId>> {
@@ -36,14 +41,31 @@ impl<T: crate::Config + pallet_session::Config> pallet_session::SessionManager<T
 	/// Updates the session index of [`pallet_session`].
 	// Instead of Some((*).expect) we could just use (*). However, we rather panic in presence of important programming errors.
 	fn new_session(new_index: SessionIndex) -> Option<Vec<T::AccountId>> {
-		debug!("PalletSessionSupport: New session {new_index}");
-		pallet_session::pallet::CurrentIndex::<T>::put(new_index);
-		Some(
-			crate::Pallet::<T>::rotate_committee_to_next_epoch()
-				.expect(
-					"PalletSessionSupport: Session should never end without current epoch validators defined. This may be caused by ShouldEndSession invalid behavior or being called before starting new session",
-				).into_iter().map(|member| member.authority_id().into()).collect::<Vec<_>>(),
-		)
+		info!("PalletSessionSupport: new_session {new_index}");
+		let new_committee = crate::Pallet::<T>::rotate_committee_to_next_epoch().expect(
+			"Session should never end without current epoch validators defined. \
+				Check ShouldEndSession implementation or if it is used before starting new session",
+		);
+		let mut keys_added: BTreeSet<T::AccountId> = BTreeSet::new();
+		for member in new_committee.iter() {
+			let account_id = member.authority_id().into();
+			if !keys_added.contains(&account_id) {
+				keys_added.insert(account_id.clone());
+				let keys = From::from(member.authority_keys());
+				let proof = sp_std::vec::Vec::new();
+				let call = pallet_session::Call::<T>::set_keys { keys, proof };
+				let res = call.dispatch_bypass_filter(RawOrigin::Signed(account_id.clone()).into());
+				match res {
+					Ok(_) => {
+						debug!("set_keys for {account_id:?}");
+					},
+					Err(e) => {
+						info!("Could not set_keys for {account_id:?}, error: {:?}", e.error)
+					},
+				}
+			}
+		}
+		Some(new_committee.into_iter().map(|member| member.authority_id().into()).collect())
 	}
 
 	fn end_session(end_index: SessionIndex) {
