+++ toolkit/committee-selection/pallet/src/tests.rs
@@ -1,6 +1,5 @@
 use crate::mock::*;
 use frame_support::{assert_ok, inherent::ProvideInherent, traits::Hooks};
-use sp_runtime::testing::UintAuthorityId;
 
 mod inherent_tests {
 	use super::*;
@@ -398,30 +397,6 @@ fn get_authority_round_robin_works() {
 		assert_eq!(
 			SessionCommitteeManagement::get_current_authority_round_robin(1),
 			Some(BOB.ids_and_keys())
-		);
-	});
-}
-
-#[test]
-fn register_session_keys_for_provided_authorities() {
-	new_test_ext().execute_with(|| {
-		System::inc_providers(&DAVE.authority_id);
-		System::inc_providers(&EVE.authority_id);
-		set_validators_directly(&[DAVE, EVE], 1).unwrap();
-		// By default, the session keys are not set for the account.
-		assert_eq!(Session::load_keys(&DAVE.authority_id), None);
-		assert_eq!(Session::load_keys(&EVE.authority_id), None);
-
-		start_session(1);
-
-		// After setting the keys, they should be stored in the session.
-		assert_eq!(
-			Session::load_keys(&DAVE.authority_id),
-			Some(SessionKeys { foo: UintAuthorityId(DAVE.authority_keys) })
-		);
-		assert_eq!(
-			Session::load_keys(&EVE.authority_id),
-			Some(SessionKeys { foo: UintAuthorityId(EVE.authority_keys) })
 		);
 	});
 }
